Move FastBQM under appropriate namespace
